requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
xgboost
lightgbm
catboost
shap
lime
interpret
Pillow
